# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enerji Yönetim Sistemi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meters:
    get:
      operationId: listMeters
      tags: [meters]
      summary: Tüm sayaçları listele
      parameters:
        - name: unitId
          in: query
          schema:
            type: integer
        - name: companyId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meter"
    post:
      operationId: createMeter
      tags: [meters]
      summary: Yeni sayaç ekle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meter"
  /meters/{id}:
    get:
      operationId: getMeter
      tags: [meters]
      summary: Sayaç detayı
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meter"
    patch:
      operationId: updateMeter
      tags: [meters]
      summary: Sayaç güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeterUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meter"
    delete:
      operationId: deleteMeter
      tags: [meters]
      summary: Sayaç sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── CONSUMPTION ──────────────────────────────────────────
...
  /risks:
    get:
      operationId: listRisks
      tags: [risks]
      summary: Risk ve fırsatlar listesi
      parameters:
        - name: unitId
          in: query
          schema:
            type: integer
        - name: companyId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiskItem"
    post:
      operationId: createRisk
      tags: [risks]
      summary: Risk/Fırsat ekle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskItem"
  /risks/{id}:
    patch:
      operationId: updateRisk
      tags: [risks]
      summary: Risk/Fırsat güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskItem"
    delete:
      operationId: deleteRisk
      tags: [risks]
      summary: Risk/Fırsat sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── SEU / ÖEK ────────────────────────────────────────────
...
components:
  schemas:
...
    Meter:
      type: object
      required: [id, name, type, location, unit]
      properties:
        id:
          type: integer
        unitId:
          type: ["integer", "null"]
        name:
          type: string
        type:
          type: string
          description: "elektrik | dogalgaz | buhar | su | diger"
        location:
          type: string
        unit:
          type: string
          description: "kWh | m3 | ton | litre"
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    MeterInput:
      type: object
      required: [name, type, location, unit]
      properties:
        unitId:
          type: integer
        name:
          type: string
        type:
          type: string
        location:
          type: string
        unit:
          type: string
        description:
          type: string
    MeterUpdate:
      type: object
      properties:
        unitId:
          type: integer
        name:
          type: string
        type:
          type: string
        location:
          type: string
        unit:
          type: string
        description:
          type: string

    # ── ConsumptionRecord ────────────────────────────────────
...
    RiskItem:
      type: object
      required: [id, type, title, probability, severity, score]
      properties:
        id:
          type: integer
        unitId:
          type: ["integer", "null"]
        type:
          type: string
          description: "risk | firsat"
        title:
          type: string
        description:
          type: ["string", "null"]
        probability:
          type: integer
          description: "1-5"
        severity:
          type: integer
          description: "1-5"
        score:
          type: integer
          description: "probability * severity"
        mitigationPlan:
          type: ["string", "null"]
        owner:
          type: ["string", "null"]
        status:
          type: string
          description: "acik | devam | kapali"
        createdAt:
          type: string
    RiskInput:
      type: object
      required: [title, probability, severity]
      properties:
        unitId:
          type: integer
        type:
          type: string
        title:
          type: string
        description:
          type: string
        probability:
          type: integer
        severity:
          type: integer
        mitigationPlan:
          type: string
        owner:
          type: string
        status:
          type: string
    RiskUpdate:
      type: object
      properties:
        unitId:
          type: integer
        type:
          type: string
        title:
          type: string
        description:
          type: string
        probability:
          type: integer
        severity:
          type: integer
        mitigationPlan:
          type: string
        owner:
          type: string
        status:
          type: string

    # ── SEU / ÖEK ────────────────────────────────────────────
...